Parallel semi-honest SPCOT
